// web_ui/src/shared/keyboard-events/keyboard.interface.ts
// INTEL CONFIDENTIAL
//
// Copyright (C) 2021 Intel Corporation
//
// This software and the related documents are Intel copyrighted materials, and your use of them is governed by
// the express license under which they were provided to you ("License"). Unless the License provides otherwise,
// you may not use, modify, copy, publish, distribute, disclose or transmit this software or the related documents
// without Intel's prior written permission.
//
// This software and the related documents are provided as is, with no express or implied warranties,
// other than those that are expressly stated in the License.

export enum KeyboardEvents {
    KeyDown = 'keydown',
    KeyUp = 'keyup',
}

export enum KeyMap {
    Z = 'Z',
    z = 'z',
    Space = ' ',
    Esc = 'Escape',
    Enter = 'Enter',
    Delete = 'Delete',
    ArrowUp = 'ArrowUp',
    ArrowDown = 'ArrowDown',
    Backspace = 'Backspace',

    Alt = 'Alt',
    Meta = 'Meta',
    Command = 'Cmd',
    Shift = 'Shift',
    Control = 'Ctrl',
}